# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Sidekeek API Service",
    "description": "Group of services provided to maintain sidekeek-client needs",
    "contact": "yourname@something.com",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/sideAPIkeek",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/editPortofolio": {
      "post": {
        "tags": ["PortofolioController"],
        "summary": "edit portofolio",
        "description": "edit portofolio",
        "operationId": "editPortofolio",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionCode",
            "in" : "formData",
            "description": "sessionCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "idPortofolio",
            "in" : "formData",
            "description": "idPortofolio",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in" : "formData",
            "description": "title",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in" : "formData",
            "description": "description",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/editPortofolio"
            }
          }
        }
      }
    },
    "/editPortofolioImg": {
      "post": {
        "tags": ["PortofolioController"],
        "summary": "edit portofolio image",
        "description": "edit sepcific portofolio image",
        "operationId": "editPortofolioImg",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionCode",
            "in" : "formData",
            "description": "sessionCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "idPortofolio",
            "in" : "formData",
            "description": "idPortofolio",
            "required": true,
            "type": "string"
          },
          {
            "name": "imgbase64",
            "in" : "formData",
            "description": "imgbase64",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/editPortofolioImg"
            }
          }
        }
      }
    },
...
    "/addNewProductDesc": {
      "post": {
        "tags": ["ProductController"],
        "summary": "add new product description",
        "description": "add new product (only description)",
        "operationId": "addNewProductDesc",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionCode",
            "in" : "formData",
            "description": "sessionCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "namaProduk",
            "in" : "formData",
            "description": "namaProduk",
            "required": true,
            "type": "string"
          },
          {
            "name": "harga",
            "in" : "formData",
            "description": "harga",
            "required": true,
            "type": "string"
          },
          {
            "name": "productDesc",
            "in" : "formData",
            "description": "productDesc",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/addNewProductDesc"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "addNewProductDesc": {
      "properties": {
        "message": {
          "example": ""
        },
        "idProduct": {
          "example": ""
        },
        "error": {
          "example": ""
        }
      }
    },
...
    "editPortofolioImg": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        }
      }
    },
    "editPortofolio": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        }
      }
    },
...
  }
}